requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
python-gitlab==1.7.0
requests==2.21.0
ruamel.yaml==0.15.87
semver==2.8.1
six==1.12.0
urllib3==1.24.1
